# src/lambdas/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "STAC API",
    "version": "1.0.0",
    "description": "This is an OpenAPI definition of the SpatioTemporal Asset Catalog API specification.",
    "contact": {
      "name": "STAC Specification",
      "url": "http://stacspec.org"
    },
    "license": {
      "name": "Apache License 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
...
  "paths": {
...
    "/collections": {
      "get": {
        "tags": [
          "Features"
        ],
        "summary": "Get the feature collections in the dataset",
        "description": "A body of Feature Collections that belong or are used together with additional links.\nRequest may not return the full set of metadata per Feature Collection.",
        "operationId": "getCollections",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Collections"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/collections/{collectionId}": {
      "get": {
        "tags": [
          "Features"
        ],
        "summary": "Describe the feature collection with Id `collectionId`",
        "description": "A single Feature Collection for the given if `collectionId`.\nRequest this endpoint to get a full list of metadata for the Feature Collection.",
        "operationId": "describeCollection",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Collection"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/conformance": {
      "get": {
        "tags": [
          "Features"
        ],
        "summary": "Get information about specifications this API conforms to",
        "description": "A list of all conformance classes specified in a standard that the\nserver conforms to.",
        "operationId": "getConformanceDeclaration",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ConformanceDeclaration"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/collections/{collectionId}/items": {
      "get": {
        "tags": [
          "Features"
        ],
        "summary": "Fetch features",
        "description": "Fetch features of the feature collection with id `collectionId`.\n\nEvery feature in a dataset belongs to a collection. A dataset may\nconsist of multiple feature collections. A feature collection is often a\ncollection of features of a similar type, based on a common schema.",
        "operationId": "getFeatures",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/bbox"
          },
          {
            "$ref": "#/components/parameters/datetime"
          },
          {
            "$ref": "#/components/parameters/query"
          },
          {
            "$ref": "#/components/parameters/filter"
          },
          {
            "$ref": "#/components/parameters/filter-lang"
          },
          {
            "$ref": "#/components/parameters/filter-crs"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Features"
          },
          "400": {
            "$ref": "#/components/responses/InvalidParameter"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/collectionId"
        }
      ],
      "post": {
        "summary": "Add a new STAC Item to a collection",
        "description": "create a new STAC Item in a specific collection",
        "operationId": "postFeature",
        "tags": [
          "Transaction"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/item"
                  },
                  {
                    "$ref": "#/components/schemas/itemCollection"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Status of the create request.",
            "headers": {
              "Location": {
                "description": "The URL of the newly added resource (i.e. path of the resource end point)",
                "schema": {
                  "type": "string",
                  "format": "url"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "A string to ensure the item has not been modified"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/item"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "default": {
            "description": "An error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collectionId}/items/{featureId}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "Fetch a single feature",
        "description": "Fetch the feature with id `featureId` in the feature collection\nwith id `collectionId`.",
        "operationId": "getFeature",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          },
          {
            "$ref": "#/components/parameters/featureId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Feature"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/collectionId"
        },
        {
          "$ref": "#/components/parameters/featureId"
        }
      ],
      "put": {
        "summary": "Update an existing feature by Id with a complete item definition",
        "description": "Use this method to update an existing feature. Requires the entire GeoJSON  description be submitted.",
        "operationId": "updateFeature",
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "description": "The request body shall contain a representation of the replacement item.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/item"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The item was replaced",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "An updated string to track changes"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "default": {
            "description": "An error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update an existing feature by Id with a partial item definition",
        "description": "Use this method to update an existing feature. Requires a GeoJSON fragment (containing the fields to be updated) be submitted.",
        "operationId": "patchFeature",
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatchOptional"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/partialItem"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Status of the update request.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "A string to ensure the item has not been modified"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "default": {
            "description": "An error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an existing feature by Id",
        "description": "Use this method to delete an existing feature.",
        "operationId": "deleteFeature",
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "responses": {
          "204": {
            "description": "The resource was deleted."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "default": {
            "description": "An error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Search STAC items with simple filtering",
        "operationId": "getItemSearch",
        "description": "Retrieve Items matching filters. Intended as a shorthand API for simple\nqueries.\n\nThis method is required to implement.\n\nIf this endpoint is implemented on a server, it is required to add a\nlink referring to this endpoint with `rel` set to `search` to the\n`links` array in `GET /`. As `GET` is the default method, the `method`\nmay not be set explicitly in the link.",
        "tags": [
          "Item Search"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/bbox"
          },
          {
            "$ref": "#/components/parameters/intersects"
          },
          {
            "$ref": "#/components/parameters/datetime"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/ids"
          },
          {
            "$ref": "#/components/parameters/collectionsArray"
          },
          {
            "$ref": "#/components/parameters/fields"
          },
          {
            "$ref": "#/components/parameters/sortby"
          },
          {
            "$ref": "#/components/parameters/query"
          },
          {
            "$ref": "#/components/parameters/filter"
          },
          {
            "$ref": "#/components/parameters/filter-lang"
          },
          {
            "$ref": "#/components/parameters/filter-crs"
          }
        ],
        "responses": {
          "200": {
            "description": "A feature collection.",
            "content": {
              "application/geo+json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/itemCollection"
                    },
                    {
                      "$ref": "#/components/schemas/context"
                    }
                  ]
                }
              },
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "summary": "Search STAC items with full-featured filtering",
        "operationId": "postItemSearch",
        "description": "Retrieve items matching filters. Intended as the standard, full-featured\nquery API.\n\nThis method is optional to implement, but recommended.\n\nIf this endpoint is implemented on a server, it is required to add a\nlink referring to this endpoint with `rel` set to `search` and `method`\nset to `POST` to the `links` array in `GET /`.",
        "tags": [
          "Item Search"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/searchBody"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A feature collection.",
            "content": {
              "application/geo+json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/itemCollection"
                    },
                    {
                      "$ref": "#/components/schemas/context"
                    }
                  ]
                }
              },
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/queryables": {
      "get": {
        "summary": "Get the JSON Schema defining the list of variable terms that can be used in CQL2 expressions",
        "operationId": "getQueryables",
        "description": "This endpoint returns a list of variable terms that can be used in CQL2 expressions. The\nprecise definition of this can be found in the OGC API - Features - Part 3: Filtering and the\nCommon Query Language (CQL2) specification.",
        "tags": [
          "Item Search"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Queryables"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/collections/{collectionId}/queryables": {
      "get": {
        "summary": "Get the JSON Schema defining the list of variable terms that can be used in CQL2 expressions",
        "operationId": "getQueryablesForCollection",
        "description": "This endpoint returns a list of variable terms that can be used in CQL2 expressions. The\nprecise definition of this can be found in the OGC API - Features - Part 3: Filtering and the\nCommon Query Language (CQL) specification.",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          }
        ],
        "tags": [
          "Features"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Queryables"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
  "components": {
...
    "responses": {
...
      "Collections": {
        "description": "The feature collections shared by this API.\n\nThe dataset is organized as one or more feature collections. This resource\nprovides information about and access to the collections.\n\nThe response contains the list of collections. For each collection, a link\nto the items in the collection (path `/collections/{collectionId}/items`,\nlink relation `items`) as well as key information about the collection.\nThis information includes:\n\n* A local identifier for the collection that is unique for the dataset;\n* A list of coordinate reference systems (CRS) in which geometries may be returned by the server. The first CRS is the default coordinate reference system (the default is always WGS 84 with axis order longitude/latitude);\n* An optional title and description for the collection;\n* An optional extent that can be used to provide an indication of the spatial and temporal extent of the collection - typically derived from the data;\n* An optional indicator about the type of the items in the collection (the default value, if the indicator is not provided, is 'feature').",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/collections"
            }
          }
        }
      },
      "Collection": {
        "description": "Information about the feature collection with id `collectionId`.\n\nThe response contains a link to the items in the collection\n(path `/collections/{collectionId}/items`, link relation `items`)\nas well as key information about the collection. This information\nincludes:\n\n* A local identifier for the collection that is unique for the dataset;\n* A list of coordinate reference systems (CRS) in which geometries may be returned by the server. The first CRS is the default coordinate reference system (the default is always WGS 84 with axis order longitude/latitude);\n* An optional title and description for the collection;\n* An optional extent that can be used to provide an indication of the spatial and temporal extent of the collection - typically derived from the data;\n* An optional indicator about the type of the items in the collection (the default value, if the indicator is not provided, is 'feature').",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/collection"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested URI was not found."
      },
      "ServerError": {
        "description": "A server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/exception"
            }
          }
        }
      },
      "ConformanceDeclaration": {
        "description": "The URIs of all conformance classes supported by the server.\n\nTo support \"generic\" clients that want to access multiple\nOGC API Features implementations - and not \"just\" a specific\nAPI / server, the server declares the conformance\nclasses it implements and conforms to.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/conformanceClasses"
            },
            "example": {
              "conformsTo": [
                "http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/core",
                "http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/oas30",
                "http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/geojson"
              ]
            }
          }
        }
      },
...
    },
    "parameters": {
      "collectionId": {
        "name": "collectionId",
        "in": "path",
        "description": "local identifier of a collection",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "datetime": {
        "name": "datetime",
        "in": "query",
        "description": "Either a date-time or an interval, open or closed. Date and time expressions\nadhere to RFC 3339. Open intervals are expressed using double-dots.\n\nExamples:\n\n* A date-time: \"2018-02-12T23:20:50Z\"\n* A closed interval: \"2018-02-12T00:00:00Z/2018-03-18T12:31:12Z\"\n* Open intervals: \"2018-02-12T00:00:00Z/..\" or \"../2018-03-18T12:31:12Z\"\n\nOnly features that have a temporal property that intersects the value of\n`datetime` are selected.\n\nIf a feature has multiple temporal properties, it is the decision of the\nserver whether only a single temporal property is used to determine\nthe extent or all relevant temporal properties.",
        "required": false,
        "schema": {
          "type": "string"
        },
        "style": "form",
        "explode": false
      },
...
      "limit": {
        "name": "limit",
        "in": "query",
        "description": "The optional limit parameter recommends the number of items that should be present in the response document.\n\nOnly items are counted that are on the first level of the collection in the response document.\nNested objects contained within the explicitly requested items shall not be counted.\n\nMinimum = 1. Maximum = 10000. Default = 10.",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 10000,
          "default": 10
        },
        "style": "form",
        "explode": false
      },
...
      "query": {
        "name": "query",
        "x-stac-api-fragment": "query",
        "in": "query",
        "description": "**Extension:** Query\n\nQuery for properties in items.",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
    }
  }
}